+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fix `get_num_unique_tiles` being unavailable for `FEEBeamCUDA`.
+- Some function comments.
+- Some clippy lints.
+
 ## [0.5.0] - 2022-08-23
 ### Added
 - `calc_jones` functions have now been renamed to "_pair" functions, which take